requirements.txt
anyascii==0.3.1
asgiref==3.5.2
beautifulsoup4==4.9.3
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coverage==6.4.4
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.0.7
django-allauth==0.51.0
django-appconf==1.0.5
django-autoslug-field==0.2.3
django-cogwheels==0.3
django-comments-xtd==2.9.7
django-compressor==4.1
django-contrib-comments==2.2.0
django-extensions==3.2.0
django-filter==21.1
django-libsass==0.9
django-modelcluster==6.0
django-modeltranslation==0.18.4
django-permissionedforms==0.1
django-recaptcha==3.0.0
django-robots==5.0
django-sass-processor==1.2.1
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
docutils==0.19
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
gcloud==0.17.0
googleapis-common-protos==1.56.4
html5lib==1.1
httplib2==0.20.4
idna==3.3
jws==0.1.3
l18n==2021.3
libsass==0.21.0
load-dotenv==0.1.0
oauth2client==3.0.0
oauthlib==3.2.0
openpyxl==3.0.10
Pillow==9.2.0
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.4.3
PyJWT==2.4.0
pyparsing==3.0.9
Pyrebase==3.0.27
pyTelegramBotAPI==4.7.0
python-dotenv==0.21.0
python-jwt==2.0.1
python3-openid==3.2.0
pytils==0.4.1
pytz==2022.2.1
rcssmin==1.1.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.7.0
rjsmin==1.2.0
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
telepath==0.2
Unidecode==1.3.4
urllib3==1.26.11
wagtail==3.0.1
wagtail-django-recaptcha==1.0
wagtail-seo==2.2.0
wagtailfontawesome==1.2.1
wagtailmenus==3.1.2
webencodings==0.5.1
whitenoise==6.2.0
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==1.4.5
xlwt==1.3.0
